Move GRPC reflection tests to GRPC module
The tests were added hin this PR: https://github.com/quarkus-qe/quarkus-test-suite/pull/1095
