+++ http/grpc/src/main/resources/application.properties
@@ -4,7 +4,10 @@ quarkus.grpc.clients.plain.use-quarkus-grpc-client=true
 # Framework doesn't use 9000 for grpc and quarkus can not detect it automatically
 quarkus.grpc.clients.plain.port=${quarkus.http.port}
 quarkus.grpc.server.use-separate-server=false
+quarkus.grpc.server.enable-reflection-service=true
 
+quarkus.grpc.clients.reflection-service.port=${quarkus.grpc.clients.plain.port}
 quarkus.grpc.clients.streaming.port=${quarkus.grpc.clients.plain.port}
 
+quarkus.grpc.clients.reflection-service.ssl.trust-store=${quarkus.grpc.clients.plain.ssl.trust-store}
 quarkus.grpc.clients.streaming.ssl.trust-store=${quarkus.grpc.clients.plain.ssl.trust-store}
