+++ http/grpc/src/test/java/io/quarkus/ts/http/grpc/OpenShiftExtensionGRPCIT.java
@@ -11,7 +11,7 @@ import io.quarkus.test.scenarios.OpenShiftScenario;
 @Tag("use-quarkus-openshift-extension")
 @OpenShiftScenario(deployment = OpenShiftDeploymentStrategy.UsingOpenShiftExtension)
 @Disabled("https://github.com/quarkus-qe/quarkus-test-framework/issues/1052+1053")
-public class OpenShiftExtensionGRPCIT implements GRPCIT, StreamingHttpIT {
+public class OpenShiftExtensionGRPCIT implements GRPCIT, StreamingHttpIT, ReflectionHttpIT {
 
     @Override
     public RestService app() {
