+++ http/grpc/src/test/java/io/quarkus/ts/http/grpc/SameServerIT.java
@@ -11,7 +11,7 @@ import io.quarkus.test.scenarios.QuarkusScenario;
 import io.quarkus.test.services.QuarkusApplication;
 
 @QuarkusScenario
-public class SameServerIT implements GRPCIT, StreamingHttpIT {
+public class SameServerIT implements GRPCIT, ReflectionHttpIT, StreamingHttpIT {
 
     @QuarkusApplication(grpc = true)
     static final GrpcService app = new GrpcService();
