+++ http/grpc/src/test/java/io/quarkus/ts/http/grpc/TLSIT.java
@@ -18,13 +18,14 @@ import io.quarkus.test.services.QuarkusApplication;
 import io.restassured.specification.RequestSpecification;
 
 @QuarkusScenario
-public class TLSIT implements GRPCIT, StreamingHttpIT {
+public class TLSIT implements GRPCIT, StreamingHttpIT, ReflectionHttpIT {
 
     private static ManagedChannel channel;
     @QuarkusApplication(grpc = true, ssl = true)
     // See https://github.com/quarkusio/quarkus/issues/38965 to learn, why we use these parameters
     static final GrpcService app = (GrpcService) new GrpcService()
             .withProperty("quarkus.grpc.clients.plain.port", "${quarkus.http.ssl-port}")
+            //            .withProperty("quarkus.grpc.clients.streaming.port", "${quarkus.http.ssl-port}")
             .withProperty("quarkus.http.ssl.certificate.files", "tls/server.pem")
             .withProperty("quarkus.http.ssl.certificate.key-files", "tls/server.key")
             .withProperty("quarkus.grpc.server.ssl.certificate", "tls/server.pem")
